test apears to work
